slack-faq-agent/requirements.txt
slack-sdk>=3.27.0
google-generativeai>=0.5.0
